fix: test 객체 변경
